[DATALAB-2091]: fixed some spark creation/reconfiguration/termination bugs

+++ infrastructure-provisioning/src/general/lib/gcp/actions_lib.py
@@ -33,6 +33,7 @@ import sys
 import time
 import traceback
 import urllib3
+import urllib.request
 import subprocess
 from Crypto.PublicKey import RSA
 from datalab.fab import *
@@ -1159,8 +1160,8 @@ class GCPActions:
                 con.sudo('sed -i \"s/^export SPARK_HOME.*/export SPARK_HOME=\/opt\/spark/\" /opt/zeppelin/conf/zeppelin-env.sh')
                 con.sudo("rm -rf /home/{}/.ensure_dir/dataengine-service_interpreter_ensure".format(ssh_user))
                 zeppelin_url = 'http://' + notebook_ip + ':8080/api/interpreter/setting/'
-                opener = urllib3.build_opener(urllib3.ProxyHandler({}))
-                req = opener.open(urllib3.Request(zeppelin_url))
+                opener = urllib.request.build_opener(urllib.request.ProxyHandler({}))
+                req = opener.open(urllib.request.Request(zeppelin_url))
                 r_text = req.read()
                 interpreter_json = json.loads(r_text)
                 interpreter_prefix = dataproc_name
@@ -1168,7 +1169,7 @@ class GCPActions:
                     if interpreter_prefix in interpreter['name']:
                         print("Interpreter with ID: {} and name: {} will be removed from zeppelin!".format(
                             interpreter['id'], interpreter['name']))
-                        request = urllib3.Request(zeppelin_url + interpreter['id'], data='')
+                        request = urllib.request.Request(zeppelin_url + interpreter['id'], data=''.encode())
                         request.get_method = lambda: 'DELETE'
                         url = opener.open(request)
                         print(url.read())
@@ -1447,7 +1448,7 @@ def remove_dataengine_kernels(notebook_name, os_user, key_path, cluster_name):
         computational_name = os.environ['computational_name'].replace('_', '-').lower()
         private = datalab.meta_lib.get_instance_private_ip_address(cluster_name, notebook_name)
         global con
-        con = datalab.fab.init_datalab_connection(private, ssh_user, key_path)
+        con = datalab.fab.init_datalab_connection(private, os_user, key_path)
         con.sudo('rm -rf /home/{}/.local/share/jupyter/kernels/*_{}'.format(os_user, cluster_name))
         if exists(con, '/home/{}/.ensure_dir/dataengine_{}_interpreter_ensured'.format(os_user, cluster_name)):
             if os.environ['notebook_multiple_clusters'] == 'true':
@@ -1464,8 +1465,8 @@ def remove_dataengine_kernels(notebook_name, os_user, key_path, cluster_name):
                 'sed -i \"s/^export SPARK_HOME.*/export SPARK_HOME=\/opt\/spark/\" /opt/zeppelin/conf/zeppelin-env.sh')
             con.sudo("rm -rf /home/{}/.ensure_dir/dataengine_interpreter_ensure".format(os_user))
             zeppelin_url = 'http://' + private + ':8080/api/interpreter/setting/'
-            opener = urllib3.build_opener(urllib3.ProxyHandler({}))
-            req = opener.open(urllib3.Request(zeppelin_url))
+            opener = urllib.request.build_opener(urllib.request.ProxyHandler({}))
+            req = opener.open(urllib.request.Request(zeppelin_url))
             r_text = req.read()
             interpreter_json = json.loads(r_text)
             interpreter_prefix = cluster_name
@@ -1473,7 +1474,7 @@ def remove_dataengine_kernels(notebook_name, os_user, key_path, cluster_name):
                 if interpreter_prefix in interpreter['name']:
                     print("Interpreter with ID: {} and name: {} will be removed from zeppelin!".format(
                         interpreter['id'], interpreter['name']))
-                    request = urllib3.Request(zeppelin_url + interpreter['id'], data='')
+                    request = urllib.request.Request(zeppelin_url + interpreter['id'], data=''.encode())
                     request.get_method = lambda: 'DELETE'
                     url = opener.open(request)
                     print(url.read())
@@ -1522,7 +1523,8 @@ def configure_dataengine_spark(cluster_name, jars_dir, cluster_dir, datalake_ena
     if os.path.exists('{0}'.format(cluster_dir)):
         subprocess.run('cp -f /tmp/{0}/notebook_spark-defaults_local.conf  {1}spark/conf/spark-defaults.conf'.format(cluster_name,
                                                                                                         cluster_dir), shell=True, check=True)
-    subprocess.run('cp -f /opt/spark/conf/core-site.xml {}spark/conf/'.format(cluster_dir), shell=True, check=True)
+    if os.path.exists('{0}'.format(cluster_dir)):
+        subprocess.run('cp -f /opt/spark/conf/core-site.xml {}spark/conf/'.format(cluster_dir), shell=True, check=True)
     if spark_configs and os.path.exists('{0}'.format(cluster_dir)):
         datalab_header = subprocess.run('cat /tmp/{0}/notebook_spark-defaults_local.conf | grep "^#"'.format(cluster_name),
                                capture_output=True, shell=True, check=True).stdout.decode('UTF-8').rstrip("\n\r")
